backend/requirements.txt
aiosqlite==0.19.0
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.8.1
anyio==4.9.0
asgiref==3.8.1
atomicwrites==1.4.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==2.0.12
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
Django==4.2.16
djangorestframework==3.15.2
filelock==3.18.0
fsspec==2025.5.1
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
huggingface-hub==0.33.0
humanfriendly==10.0
humanize==4.12.3
idna==3.10
iniconfig==2.1.0
netaddr==1.3.0
numpy==2.2.6
openvpn-api==0.3.0
openvpn-status==0.2.2
packaging==25.0
pandas==2.3.0
pefile==2023.2.7
pillow==10.0.0
pluggy==1.0.0.dev0
psycopg2-binary==2.9.10
py==1.11.0
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.11.5
pydantic_core==2.33.2
pyinstaller==6.14.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.1.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.6
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.1
requests==2.26.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
Telethon==1.34.0
tokenizers==0.21.1
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
